requirements.txt
python-dotenv
Requests
websockets
ccxt
pandas
ccxt
matplotlib
numpy